requirements.txt
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
Flask==2.3.2
google-ads==21.2.0
google-api-core==2.11.0
google-auth==2.20.0
google-auth-oauthlib==0.8.0
googleads==38.0.0
googleapis-common-protos==1.59.1
grpcio==1.56.0
grpcio-status==1.56.0
idna==3.4
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.3
oauthlib==3.2.2
platformdirs==3.9.1
proto-plus==1.22.3
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rsa==4.9
six==1.16.0
urllib3==1.26.16
Werkzeug==2.3.6
xmltodict==0.13.0
zeep==4.2.1